+++ src/tasks.ts
@@ -1,50 +1,189 @@
 import * as vscode from 'vscode';
 import * as path from 'path';
 import * as fs from 'fs';
+import * as os from 'os';
 import * as cp from 'child_process';
 
-import {
-  getCustomBuildFile,
-  getIsAutoDetectionEnabled,
-  getTasksArgs
-} from './config';
+import { getIsAutoDetectionEnabled } from './config';
 
-type StringMap = { [s: string]: string };
+type GradleTask = {
+  path: string;
+  description: string;
+  name: string;
+  group: string;
+  project: string;
+};
+
+type GradleProject = {
+  name: string;
+  path: string;
+  parent: string;
+  tasks: GradleTask[];
+  buildFile: string;
+};
+
+export interface GradleTaskDefinition extends vscode.TaskDefinition {
+  script: string;
+  description: string;
+  group: string;
+  project: string;
+  rootProject: string;
+  buildFile: string;
+}
+
 let autoDetectOverride: boolean = false;
-let cachedTasks: Promise<vscode.Task[]> | undefined = undefined;
+let cachedTasks: Promise<vscode.Task[]> | undefined;
+let refreshProcess: cp.ChildProcessWithoutNullStreams | undefined;
 
 export function enableTaskDetection() {
   autoDetectOverride = true;
 }
 
-export interface GradleTaskDefinition extends vscode.TaskDefinition {
-  script: string;
-  fileName: string;
-  description: string;
+export function killRefreshProcess() {
+  if (refreshProcess) {
+    refreshProcess.kill();
+  }
+}
+
+async function isGradleProject(
+  folder: vscode.WorkspaceFolder
+): Promise<boolean> {
+  const defaultBuildFileGlob = '{*.gradle,*.gradle.kts}';
+  const relativePattern = new vscode.RelativePattern(
+    folder,
+    defaultBuildFileGlob
+  );
+  const files = await vscode.workspace.findFiles(relativePattern);
+  return files.length > 0;
 }
 
 export class GradleTaskProvider implements vscode.TaskProvider {
   constructor(
     readonly statusBarItem: vscode.StatusBarItem,
-    readonly outputChannel: vscode.OutputChannel
+    readonly outputChannel: vscode.OutputChannel,
+    readonly context: vscode.ExtensionContext
   ) {}
 
   async provideTasks() {
     try {
-      return await provideGradleTasks(this.statusBarItem, this.outputChannel);
+      return await this.provideGradleTasks();
     } catch (e) {
-      this.outputChannel.append(`Error providing gradle tasks: ${e.message}\n`);
-      this.outputChannel.show();
+      const message = `Error providing gradle tasks: ${e.message}`;
+      console.error(message);
+      this.outputChannel.appendLine(message);
     }
   }
 
-  // TODO: write tests that cover the case where auto-discover tasks is
-  // switched off, but we can still run a task by setting the task definition
-  // within tasks.json
+  // TODO
   public async resolveTask(
     _task: vscode.Task
   ): Promise<vscode.Task | undefined> {
     return undefined;
+  }
+
+  private provideGradleTasks(): Promise<vscode.Task[]> {
+    if (!cachedTasks) {
+      cachedTasks = this.detectGradleTasks();
+    }
+    return cachedTasks;
+  }
+
+  private async detectGradleTasks(): Promise<vscode.Task[]> {
+    const emptyTasks: vscode.Task[] = [];
+    const allTasks: vscode.Task[] = [];
+    const folders = vscode.workspace.workspaceFolders;
+    if (!folders) {
+      return emptyTasks;
+    }
+    try {
+      for (const folder of folders) {
+        if (autoDetectOverride || getIsAutoDetectionEnabled(folder)) {
+          if (await isGradleProject(folder)) {
+            const tasks = await this.provideGradleTasksForFolder(folder);
+            allTasks.push(...tasks);
+          }
+        }
+      }
+      return allTasks;
+    } catch (error) {
+      return Promise.reject(error);
+    }
+  }
+
+  private async provideGradleTasksForFolder(
+    folder: vscode.WorkspaceFolder
+  ): Promise<vscode.Task[]> {
+    const emptyTasks: vscode.Task[] = [];
+
+    // const folder = vscode.workspace.getWorkspaceFolder(gradleBuildFileUri);
+    // if (!folder) {
+    //   return emptyTasks;
+    // }
+    const command = await getGradleWrapperCommandFromPath(folder.uri.fsPath);
+    if (!command) {
+      return emptyTasks;
+    }
+    const gradleProjects:
+      | GradleProject[]
+      | undefined = await this.getGradleProjects(folder);
+    if (!gradleProjects || !gradleProjects.length) {
+      return emptyTasks;
+    }
+
+    const rootProject: GradleProject | undefined = gradleProjects.find(
+      project => !project.parent
+    );
+    if (!rootProject) {
+      return emptyTasks;
+    }
+
+    const tasks: vscode.Task[] = [];
+
+    for (const gradleProject of gradleProjects) {
+      for (const gradleTask of gradleProject.tasks) {
+        tasks.push(
+          await createTask(
+            gradleTask,
+            folder!,
+            command,
+            rootProject.name,
+            vscode.Uri.file(gradleProject.buildFile)
+          )
+        );
+      }
+    }
+
+    return tasks;
+  }
+
+  private async getProjectsFromGradle(
+    folder: vscode.WorkspaceFolder
+  ): Promise<string> {
+    this.statusBarItem.text = '$(sync~spin) Refreshing gradle tasks';
+    this.statusBarItem.show();
+    const tempDir = await fs.promises.mkdtemp(
+      path.join(os.tmpdir(), 'vscode-gradle-')
+    );
+    const tempFile = path.join(tempDir, 'tasks.json');
+    const command = getTasksScriptCommand();
+    const cwd = this.context.asAbsolutePath('lib');
+    const args = [folder.uri.fsPath, tempFile];
+    return spawn(command, args, { cwd }, this.outputChannel)
+      .then(() => fs.promises.readFile(tempFile, 'utf8'))
+      .finally(async () => {
+        this.statusBarItem.hide();
+        if (await exists(tempFile)) {
+          await fs.promises.unlink(tempFile);
+        }
+        await fs.promises.rmdir(tempDir);
+      });
+  }
+
+  private async getGradleProjects(
+    folder: vscode.WorkspaceFolder
+  ): Promise<GradleProject[] | undefined> {
+    const stdout = await this.getProjectsFromGradle(folder);
+    return parseGradleProjects(stdout);
   }
 }
 
@@ -75,169 +214,46 @@ export async function getGradleWrapperCommandFromPath(
   }
 }
 
-export async function getBuildFilePaths(
-  uri: vscode.Uri
-): Promise<vscode.Uri[]> {
-  const customBuildFile = getCustomBuildFile(uri);
-  const customBuildFileGlob = customBuildFile && `{${customBuildFile}}`;
-  const defaultBuildFileGlob = '{build.gradle,build.gradle.kts}';
-  const buildFileGlob = customBuildFileGlob || defaultBuildFileGlob;
-  const relativePattern = new vscode.RelativePattern(uri.fsPath, buildFileGlob);
-  const paths = await vscode.workspace.findFiles(relativePattern);
-  return paths;
-}
-
-async function detectGradleTasks(
-  statusBarItem: vscode.StatusBarItem,
-  outputChannel: vscode.OutputChannel
-): Promise<vscode.Task[]> {
-  const emptyTasks: vscode.Task[] = [];
-  const allTasks: vscode.Task[] = [];
-  const visitedBuildGradleFiles: Set<string> = new Set();
-
-  const folders = vscode.workspace.workspaceFolders;
-  if (!folders) {
-    return emptyTasks;
+export function getTasksScriptCommand(): string {
+  const platform = process.platform;
+  if (platform === 'win32') {
+    return '.\\gradle-tasks.bat';
+  } else if (platform === 'linux' || platform === 'darwin') {
+    return './gradle-tasks';
+  } else {
+    throw new Error('Unsupported platform');
   }
-  try {
-    for (const folder of folders) {
-      if (autoDetectOverride || getIsAutoDetectionEnabled(folder)) {
-        const paths = await getBuildFilePaths(folder.uri);
-        for (const path of paths) {
-          if (!visitedBuildGradleFiles.has(path.fsPath)) {
-            const tasks = await provideGradleTasksForFolder(
-              path,
-              statusBarItem,
-              outputChannel
-            );
-            visitedBuildGradleFiles.add(path.fsPath);
-            allTasks.push(...tasks);
-          }
-        }
-      }
-    }
-    return allTasks;
-  } catch (error) {
-    return Promise.reject(error);
-  }
-}
-
-export function provideGradleTasks(
-  statusBarItem: vscode.StatusBarItem,
-  outputChannel: vscode.OutputChannel
-): Promise<vscode.Task[]> {
-  if (!cachedTasks) {
-    cachedTasks = detectGradleTasks(statusBarItem, outputChannel);
-  }
-  return cachedTasks;
-}
-
-async function provideGradleTasksForFolder(
-  gradleBuildFileUri: vscode.Uri,
-  statusBarItem: vscode.StatusBarItem,
-  outputChannel: vscode.OutputChannel
-): Promise<vscode.Task[]> {
-  const emptyTasks: vscode.Task[] = [];
-
-  const folder = vscode.workspace.getWorkspaceFolder(gradleBuildFileUri);
-  if (!folder) {
-    return emptyTasks;
-  }
-  const command = await getGradleWrapperCommandFromPath(folder.uri.fsPath);
-  if (!command) {
-    return emptyTasks;
-  }
-  const tasksMap: StringMap | undefined = await getTasks(
-    command,
-    folder,
-    statusBarItem,
-    outputChannel
-  );
-  if (!tasksMap) {
-    return emptyTasks;
-  }
-
-  const sortedKeys = Object.keys(tasksMap).sort((a, b) => a.localeCompare(b));
-
-  const tasks: vscode.Task[] = [];
-  for (const taskName of sortedKeys) {
-    tasks.push(
-      await createTask(
-        taskName,
-        tasksMap[taskName],
-        folder!,
-        gradleBuildFileUri,
-        command
-      )
-    );
-  }
-  return tasks;
-}
-
-export function getTaskName(task: string, relativePath: string | undefined) {
-  if (relativePath && relativePath.length) {
-    return `${task} - ${relativePath.substring(0, relativePath.length - 1)}`;
-  }
-  return task;
 }
 
 export async function createTask(
-  taskName: string,
-  taskDescription: string,
+  gradleTask: GradleTask,
   folder: vscode.WorkspaceFolder,
-  gradleBuildFileUri: vscode.Uri,
-  command: string
+  command: string,
+  rootProject: string,
+  buildFile: vscode.Uri
 ): Promise<vscode.Task> {
   function getCommandLine(task: string): string {
-    const args: string[] = [];
-    args.push(task);
-    const customBuildFile = getCustomBuildFile(folder.uri);
-    if (customBuildFile) {
-      args.push('--build-file', customBuildFile);
-    }
-    return `"${command}" ${args.join(' ')}`;
+    return `"${command}" ${task}`;
   }
 
-  function getRelativePath(
-    folder: vscode.WorkspaceFolder,
-    gradleBuildFileUri: vscode.Uri
-  ): string {
-    return path.relative(
-      folder.uri.fsPath,
-      path.dirname(gradleBuildFileUri.fsPath)
-    );
-  }
-
-  let fileName: string = path.basename(gradleBuildFileUri.fsPath);
-  if (taskName.includes(':')) {
-    const [subProjectName] = taskName.split(':');
-    const subProjectPath = path.join(folder.uri.fsPath, subProjectName);
-    const folderUri = vscode.Uri.file(subProjectPath);
-    const uri = (await getBuildFilePaths(folderUri))[0];
-    fileName = path.basename(uri.fsPath);
-  }
+  const friendlyTaskName = gradleTask.path.replace(/^:/, '');
   const definition: GradleTaskDefinition = {
     type: 'gradle',
-    script: taskName,
-    fileName,
-    description: taskDescription
+    script: friendlyTaskName,
+    description: gradleTask.description,
+    group: (gradleTask.group || 'other').toLowerCase(),
+    project: gradleTask.project,
+    buildFile: buildFile.fsPath,
+    rootProject
   };
-
-  const relativeBuildGradle = getRelativePath(folder, gradleBuildFileUri);
-  if (relativeBuildGradle.length) {
-    definition.path = relativeBuildGradle;
-  }
-  const normalizedTaskName = getTaskName(
-    definition.script,
-    relativeBuildGradle
-  );
-  const cwd = path.dirname(gradleBuildFileUri.fsPath);
+  const cwd = folder.uri.fsPath;
+  const cmd = getCommandLine(definition.script);
   const task = new vscode.Task(
     definition,
     folder,
-    normalizedTaskName,
+    definition.script,
     'gradle',
-    new vscode.ShellExecution(getCommandLine(taskName), { cwd }),
+    new vscode.ShellExecution(cmd, { cwd }),
     ['$gradle']
   );
   task.presentationOptions = {
@@ -248,7 +264,7 @@ export async function createTask(
   return task;
 }
 
-export async function hasGradleBuildFile(): Promise<boolean> {
+export async function hasGradleProject(): Promise<boolean> {
   const folders = vscode.workspace.workspaceFolders;
   if (!folders) {
     return false;
@@ -257,28 +273,7 @@ export async function hasGradleBuildFile(): Promise<boolean> {
     if (folder.uri.scheme !== 'file') {
       continue;
     }
-    const customBuildFile = getCustomBuildFile(folder.uri);
-    if (customBuildFile) {
-      const customBuildFilePath = path.join(folder.uri.fsPath, customBuildFile);
-      if (await exists(customBuildFilePath)) {
-        return true;
-      } else {
-        // If custom build filename is set then don't check for default build files
-        continue;
-      }
-    }
-    const defaultGroovyBuildFilePath = path.join(
-      folder.uri.fsPath,
-      'build.gradle'
-    );
-    const defaultKotlinBuildFilePath = path.join(
-      folder.uri.fsPath,
-      'build.gradle.kts'
-    );
-    if (
-      (await exists(defaultGroovyBuildFilePath)) ||
-      (await exists(defaultKotlinBuildFilePath))
-    ) {
+    if (await isGradleProject(folder)) {
       return true;
     }
   }
@@ -293,95 +288,46 @@ async function exists(file: string): Promise<boolean> {
   });
 }
 
-const TASK_REGEX: RegExp = /$\s*(([^:\s]*:)?([a-z]+[A-Z0-9]?[a-z0-9]*[A-Za-z0-9]*))\b(\s-\s(.*))?/gm;
-
-export function parseGradleTasks(buffer: Buffer | string): StringMap {
-  const tasks: StringMap = {};
-  let match: RegExpExecArray | null = null;
-  while ((match = TASK_REGEX.exec(buffer.toString())) !== null) {
-    const [, name, , , , description] = match;
-    tasks[name] = description;
+export function parseGradleProjects(buffer: Buffer | string): GradleProject[] {
+  const input = buffer.toString();
+  try {
+    return JSON.parse(input);
+  } catch (e) {
+    throw new Error(`${e.message} - input: ${input}`);
   }
-  return tasks;
-}
-
-function getBuffersAsString(buffers: Buffer[]): string {
-  return Buffer.concat(buffers)
-    .toString('utf8')
-    .trim();
 }
 
 function debugCommand(
   command: string,
-  args: ReadonlyArray<string> = [],
+  args: ReadonlyArray<string>,
+  options: cp.SpawnOptionsWithoutStdio,
   outputChannel: vscode.OutputChannel
 ) {
-  outputChannel.append(`Executing: ${command} ${args.join(' ')}\n`);
+  outputChannel.appendLine(`Executing: ${command} ${args.join(' ')}`);
 }
 
 export function spawn(
   command: string,
   args: ReadonlyArray<string> = [],
-  options: cp.ExecOptions = {},
-  outputChannel?: vscode.OutputChannel
+  options: cp.SpawnOptionsWithoutStdio = {},
+  outputChannel: vscode.OutputChannel
 ): Promise<string> {
-  if (outputChannel) {
-    debugCommand(command, args, outputChannel);
-  }
+  debugCommand(command, args, options, outputChannel);
   return new Promise((resolve, reject) => {
-    const stdoutBuffers: Buffer[] = [];
-    const stderrBuffers: Buffer[] = [];
-    const child = cp.spawn(command, args, options);
-    child.stdout.on('data', (buffer: Buffer) => stdoutBuffers.push(buffer));
-    child.stderr.on('data', (buffer: Buffer) => stderrBuffers.push(buffer));
-    child.on('error', err => {
-      reject(err);
-    });
-    child.on('exit', (code: number) => {
+    refreshProcess = cp.spawn(command, args, options);
+    refreshProcess.stdout.on('data', (buffer: Buffer) =>
+      outputChannel.append(buffer.toString())
+    );
+    refreshProcess.stderr.on('data', (buffer: Buffer) =>
+      outputChannel.append(buffer.toString())
+    );
+    refreshProcess.on('error', reject);
+    refreshProcess.on('exit', (code: number) => {
       if (code === 0) {
-        resolve(getBuffersAsString(stdoutBuffers));
+        resolve();
       } else {
-        reject(new Error(getBuffersAsString(stderrBuffers)));
+        reject();
       }
     });
   });
-}
-
-function getTasksFromGradle(
-  command: string,
-  folder: vscode.WorkspaceFolder,
-  statusBarItem: vscode.StatusBarItem,
-  outputChannel: vscode.OutputChannel
-): Promise<string> {
-  statusBarItem.text = '$(sync~spin) Refreshing gradle tasks';
-  statusBarItem.show();
-
-  const args = ['--quiet', 'tasks'];
-  const tasksArgs = getTasksArgs(folder);
-  if (tasksArgs) {
-    args.push(tasksArgs);
-  }
-  const customBuildFile = getCustomBuildFile(folder.uri);
-  if (customBuildFile) {
-    args.push('--build-file', customBuildFile);
-  }
-  const { fsPath: cwd } = folder.uri;
-  return spawn(command, args, { cwd }, outputChannel).finally(() => {
-    statusBarItem.hide();
-  });
-}
-
-async function getTasks(
-  command: string,
-  folder: vscode.WorkspaceFolder,
-  statusBarItem: vscode.StatusBarItem,
-  outputChannel: vscode.OutputChannel
-): Promise<StringMap | undefined> {
-  const stdout = await getTasksFromGradle(
-    command,
-    folder,
-    statusBarItem,
-    outputChannel
-  );
-  return parseGradleTasks(stdout);
 }